pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastx402"
version = "0.1.0"
description = "FastAPI integration for x402 HTTP-native payments"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "fastx402 contributors"}
]
keywords = ["fastapi", "x402", "payments", "ethereum", "web3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "eth-account>=0.9.0",
    "eth-utils>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "uvicorn[standard]>=0.24.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]
waas = [
    "privy>=0.1.0",  # Privy Python SDK for embedded wallets
]
ws = [
    "websockets>=12.0",  # WebSocket server for x402instant communication
]

[tool.setuptools.packages.find]
where = ["."]
include = ["fastx402*"]

